Rename patchmodal to patchrestartmodal
