+++ src/pages/patch/index.ts
@@ -2,6 +2,6 @@ import { ActionButtons } from "pages/patch/ActionButtons";
 import { Metadata } from "pages/patch/Metadata";
 import { PatchTabs } from "pages/patch/PatchTabs";
 import { BuildVariants } from "pages/patch/BuildVariants";
-import { PatchModal } from "pages/patch/PatchModal";
+import { PatchRestartModal } from "pages/patch/PatchRestartModal";
 
-export { ActionButtons, Metadata, PatchTabs, BuildVariants, PatchModal };
+export { ActionButtons, Metadata, PatchTabs, BuildVariants, PatchRestartModal };
