+++ src/pages/patch/patchRestartModal/index.ts
